requirements.txt
# Core
numpy>=1.26
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
joblib>=1.3
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
plotly>=5.22
seaborn>=0.13

# Models / tuning
xgboost>=2.0
lightgbm>=4.3
optuna>=3.6

# Explainability & tracking
shap>=0.45
mlflow>=2.13

# API & App
fastapi>=0.111
pydantic>=2.7
uvicorn>=0.30
streamlit>=1.36

# Dev
pytest>=8.0
pytest-cov>=4.1
black>=24.3.0
ruff>=0.5.0
